+++ src/CzechHolidays/CzechHolidayDateComparer.cs
@@ -1,9 +1,0 @@
-﻿namespace CzechHolidays;
-
-internal sealed class CzechHolidayDateComparer : IComparer<CzechHolidayDate>
-{
-    internal CzechHolidayDateComparer() { }
-
-    public int Compare(CzechHolidayDate x, CzechHolidayDate y) => x.Date.CompareTo(y.Date);
-}
-